Create Power Of Number

+++ calculator/SimpleAlgoritms.cs
@@ -16,9 +16,9 @@ namespace calculator
             }
             return sum == number ? true : false;
         }
-        public int PowerOfNumber(int number, int power)
+        public double PowerOfNumber(double number, int power)
         {
-            int result = 1;
+            double result = 1;
             for (int i = 0; i < power; i++) result *= number;
             return result;
         }
